agents/requirements.txt
# ─── Trinetra Agents: Shared Requirements ───
# All 13 agents share this single requirements file.

# ── Redis Pub/Sub (replaces Kafka) ──
redis>=5.0.0

# ── HTTP Client (Backend API + External APIs) ──
requests>=2.31.0

# ── Document Parsing (Doc Intelligence Agent) ──
pdfplumber>=0.11.0
pytesseract>=0.3.13
Pillow>=11.1.0

# ── Data Manipulation ──
pandas>=2.2.0
numpy>=1.26.4

# ── GST Agent: Graph Theory ──
networkx>=3.2.1

# ── Web Agent: RAG Pipeline ──
weaviate-client>=3.26.2
sentence-transformers>=2.3.1
vaderSentiment>=3.3.2

# ── PD Agent: Speech-to-Text + LLM ──
openai-whisper>=20240930
groq>=0.4.2

# ── Risk Agent: ML + Explainability ──
xgboost>=2.0.3
lightgbm>=4.3.0
shap>=0.44.0
lime>=0.2.0.1
scikit-learn>=1.4.0

# ── CAM Agent: Word Document Generation ──
python-docx>=1.1.0

# ── Utilities ──
python-dotenv>=1.0.1
boto3>=1.34.25
flask>=3.0.0
